+++ src/__test__/strategy.test.ts
@@ -1,28 +1,28 @@
 import {createRegion} from '..';
 import {delayLoop} from '../util/delayLoop';
 
-const fn1 = async () => {
+const asyncFunc1 = async () => {
     await delayLoop();
     return 'a';
 };
 
-const fn2 = async () => {
+const asyncFunc2 = async () => {
     await delayLoop(2);
     return 'b';
 };
 
 describe('strategy', () => {
     test('fn', async () => {
-        const result1 = await Promise.race([fn1(), fn2()]);
-        const result2 = await Promise.race([fn2(), fn1()]);
+        const result1 = await Promise.race([asyncFunc1(), asyncFunc2()]);
+        const result2 = await Promise.race([asyncFunc2(), asyncFunc1()]);
         expect(result1).toBe('a');
         expect(result2).toBe('a');
     });
 
     test('acceptSequenced 1', async () => {
         const region = createRegion();
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         await load1();
         expect(region.getValue()).toBe('a');
         await load2();
@@ -31,8 +31,8 @@ describe('strategy', () => {
 
     test('acceptSequenced 2', async () => {
         const region = createRegion();
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load1(), load2()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('a');
@@ -42,8 +42,8 @@ describe('strategy', () => {
 
     test('acceptSequenced 3', async () => {
         const region = createRegion();
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load2(), load1()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('a');
@@ -53,8 +53,8 @@ describe('strategy', () => {
 
     test('acceptLatest 1', async () => {
         const region = createRegion(undefined, {strategy: 'acceptLatest'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         await load1();
         expect(region.getValue()).toBe('a');
         await load2();
@@ -63,8 +63,8 @@ describe('strategy', () => {
 
     test('acceptLatest 2', async () => {
         const region = createRegion(undefined, {strategy: 'acceptLatest'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load1(), load2()];
         await Promise.race(promises);
         expect(region.getValue()).toBe(undefined);
@@ -74,8 +74,8 @@ describe('strategy', () => {
 
     test('acceptLatest 3', async () => {
         const region = createRegion(undefined, {strategy: 'acceptLatest'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load2(), load1()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('a');
@@ -85,8 +85,8 @@ describe('strategy', () => {
 
     test('acceptEvery 1', async () => {
         const region = createRegion(undefined, {strategy: 'acceptEvery'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         await load1();
         expect(region.getValue()).toBe('a');
         await load2();
@@ -95,8 +95,8 @@ describe('strategy', () => {
 
     test('acceptEvery 2', async () => {
         const region = createRegion(undefined, {strategy: 'acceptEvery'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load1(), load2()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('a');
@@ -106,8 +106,8 @@ describe('strategy', () => {
 
     test('acceptEvery 3', async () => {
         const region = createRegion(undefined, {strategy: 'acceptEvery'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load2(), load1()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('a');
@@ -117,8 +117,8 @@ describe('strategy', () => {
 
     test('acceptFirst 1', async () => {
         const region = createRegion(undefined, {strategy: 'acceptFirst'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         await load1();
         expect(region.getValue()).toBe('a');
         await load2();
@@ -127,8 +127,8 @@ describe('strategy', () => {
 
     test('acceptFirst 2', async () => {
         const region = createRegion(undefined, {strategy: 'acceptFirst'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load1(), load2()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('a');
@@ -138,8 +138,8 @@ describe('strategy', () => {
 
     test('acceptFirst 3', async () => {
         const region = createRegion(undefined, {strategy: 'acceptFirst'});
-        const load1 = region.loadBy(fn1);
-        const load2 = region.loadBy(fn2);
+        const load1 = region.loadBy(asyncFunc1);
+        const load2 = region.loadBy(asyncFunc2);
         const promises = [load2(), load1()];
         await Promise.race(promises);
         expect(region.getValue()).toBe('b');